Create README.md
Hey there millie! I say if things keep going the way they are, by the time your 30 youre definitely going to be looking for more and more excitement. Now, if you ever find the thing to fulfill that excitement?...that would be the interesting thing to see in my opinion. By this point youve experienced the humdrum lifestyle of motherhood, possibly marriage and maybe even divorce? Either way i dont think it would matter to you in your mind, at this point youre realizing youve hit 30 and youre going to do WHATEVER it takes to find that missing thing in your life. Will it be the excitement of flirting with a random stranger, or maybe even the adrenaline rush of continuously having a secret affair behind your significant others back;) All you know is that when you look at yourself in the mirror, its not sadness but determination you feel every morning, determination to prove that youre still that hot young woman you once were in high school...you may not be completely pleased with what you are staring at, but hey that happens to us all as we get older right? The one thing in life that is undefeated is good old father time. However all that realization does to you, is make you even more anxious to find that void filler in your life, even if it means risking the current droll life youve built up till this point! Now maybe one day youre out running some errands by yourself without the husband, without the kids, and thats when you run into me. Currently 28 so by this point in your life 38yrs old having that slight age gap between us. All i see is a something....different, something that catches my eye. I recognize your beauty immediately, and whatever new life -excitement-goal vibes you are clearly giving off. I also find myself wondering where my life has gone till this point, being from the west coast and now married to an iowa woman with a couple of kids. All i can tell myself in that moment though is that, "I need to tell this young, stunning woman, hi at the very least. Shoot my shot" idk what it is about you that has now all of a sudden made me have this uncontrollable desire to just, make you happy, make myself happy again, and start this new forbidden fling... all i could say to you, at this moment, is that you are the most stunning woman ive seen in so long and id love to take you out to dinner. No...fuck that, id want to just skip that and convince myself, and you, to just give in to this spontaneous moment, as we decide to leave right then and there from the location we are at, and head to a hotel for some fun. The hardest part now however, is being able to control myself long enough to get to the hotel, but all we want to do is say "fuck it" and enjoy some adult rated fun right there in the car;) Thus starting a new, forbidden, fling in our lives..fulfilling the desires you had of knowing your that sexy young and fun woman still and fulfilling my desires of wanting to feel young and care free again as we continue something secretive, risky, yet that much more exciting;)
-Rick